+++ htmlfmt.php
@@ -2,6 +2,9 @@
 $LlCwrs="";
 $LlGwers="";
 $LlTeitl="";
+$LlCynTeitl="";
+$LlNesaf="";
+$LlCynModiwl="";
 $LlPlygellSain="";
 $LlNodwch="";
 $LlCyfarwyddo="";
@@ -29,10 +32,13 @@ $DFfeil1 = explode("\n", file_get_contents("./". $LlTestun) );
 //---------------------------------------
 //---------------------------------------
 function htmlfmtinit(){
-global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlFfeilPrev; global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun; global $LlDidoli; global $LlPlygellSain;
+global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlFfeilPrev; global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun; global $LlDidoli; global $LlPlygellSain; global $LlCynTeitl; global $LlCynModiwl; global $LlNesaf;
 $LlCwrs="";
 $LlGwers="";
 $LlTeitl="";
+$LlCynTeitl="";
+$LlNesaf="";
+$LlCynModiwl="";
 $LlPlygellSain="";
 $LlNodwch="";
 $LlCyfarwyddo="";
@@ -52,7 +58,7 @@ $LlEn="";
 //---------------------------------------
 //---------------------------------------
 function htmlfmtsettings($ll1a){
-global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlFfeil;  global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun; global $LlDidoli; global $LlPlygellSain;
+global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlFfeil;  global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun; global $LlDidoli; global $LlPlygellSain; global $LlCynTeitl; global $LlCynModiwl; global $LlNesaf;
 $bfound=false;
       //____$d1a = explode("=", substr($ll1a,1));
       $d1a = explode("=", substr($ll1a,0));
@@ -68,6 +74,15 @@ $bfound=false;
         $bfound=true;
       }else if($d1a[0]=="teitl"){
         $LlTeitl=$d1a[1];
+        $bfound=true;
+      }else if($d1a[0]=="cynteitl"){
+        $LlCynTeitl=$d1a[1];
+        $bfound=true;
+      }else if($d1a[0]=="nesaf"){
+        $LlNesaf=$d1a[1];
+        $bfound=true;
+      }else if($d1a[0]=="cynmodiwl"){
+        $LlCynModiwl=$d1a[1];
         $bfound=true;
       }else if($d1a[0]=="cwrs"){
         $LlCwrs.=$d1a[1]. "/";